+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased - 123
 
 - Minor UI improvements
+- Fix new design not being usable on devices with square screens
 
 ## v2.23.2 - 122